implement delete 2 R-trees built on the same file

+++ Algebras/RTree/RTreeAlgebra.h
@@ -1591,6 +1591,7 @@ Opens an existing R-tree.
     void SwitchHeader(R_Tree<dim,LeafInfo>*);
     void MergeRtree();
     SmiRecordId Record_Path_Id(){return header.path_rec_id;}
+    int GetShare(){return header.share;}
 ////////////////////////////////////////////////////////////////////////////
     void Clone(R_Tree<dim,LeafInfo>*);
     SmiRecordId DFVisit_Rtree(R_Tree<dim,LeafInfo>*,R_TreeNode<dim,LeafInfo>*);
@@ -1851,20 +1852,21 @@ The record file of the R-Tree.
       //new record
       SmiRecordId second_head_id;    //two rtrees on the same file
       SmiRecordId path_rec_id;  //record update path for new coverage
+      int share;
 
       Header() :
         headerRecordId( 0 ), rootRecordId( 0 ),
         minLeafEntries( 0 ), maxLeafEntries( 0 ),
         minInternalEntries( 0 ), maxInternalEntries( 0 ),
         nodeCount( 0 ), entryCount( 0 ), height( 0 ),
-        second_head_id(0), path_rec_id(0)
+        second_head_id(0), path_rec_id(0), share(0)
         {}
       Header( SmiRecordId headerRecordId, SmiRecordId rootRecordId = 0,
               int minEntries = 0, int maxEntries = 0,
               int minInternalEntries = 0, int maxInternalEntries = 0,
               int nodeCount = 0, int entryCount = 0,
               int nodeSize = 0, int height = 0,
-              SmiRecordId s_h_i = 0, SmiRecordId p_r_i = 0) :
+              SmiRecordId s_h_i = 0, SmiRecordId p_r_i = 0, int s = 0) :
         headerRecordId( headerRecordId ),
         rootRecordId( rootRecordId ),
         minLeafEntries( minLeafEntries ),
@@ -1874,7 +1876,7 @@ The record file of the R-Tree.
         nodeCount( nodeCount ),
         entryCount( entryCount ),
         height( height ),
-        second_head_id(s_h_i), path_rec_id(p_r_i)
+        second_head_id(s_h_i), path_rec_id(p_r_i), share(s)
         {}
     } header;
 
@@ -2369,6 +2371,8 @@ nodeIdCounter( 0 )
 template <unsigned dim, class LeafInfo>
 R_Tree<dim, LeafInfo>::~R_Tree()
 {
+//    cout<<"~R_Tree() header share "<<GetShare()<<endl;
+    header.share--;
     if( file->IsOpen() )
    {
      if( nodePtr != NULL )
@@ -2380,10 +2384,9 @@ R_Tree<dim, LeafInfo>::~R_Tree()
        file->Close();
    }
 
-   if( fileOwner ){
-//      cout<<"second_head_id "<<header.second_head_id<<endl;
+   if( fileOwner)
       delete file;
-  }
+
 }
 
 /*
@@ -3783,6 +3786,7 @@ void R_Tree<dim, LeafInfo>::MergeRtree()
       header.nodeCount += tree1_node_count;
       header.rootRecordId = root1_id;
       header.path_rec_id = path_rec_id;
+      header.share = 2;
       WriteHeader();
 
   }else if(temp_head.height == header.height){
@@ -3838,6 +3842,7 @@ void R_Tree<dim, LeafInfo>::MergeRtree()
         header.path_rec_id = path_rec_id;
                /////////////
         path[0] = new_node_rec_id;
+        header.share = 2;
         WriteHeader();
 //    cout<<"header "<<header.headerRecordId<<"root id "<<RootRecordId()<<endl;
 /*        ReadHeader();
@@ -3947,6 +3952,7 @@ void R_Tree<dim, LeafInfo>::MergeRtree()
       header.rootRecordId = adr2;
       header.path_rec_id = path_rec_id;
 //      cout<<"write path_rec_id "<<header.path_rec_id<<endl;
+      header.share = 2;
       WriteHeader();
 
 //      ReadHeader();
@@ -4375,8 +4381,14 @@ void DeleteRTree( const ListExpr typeInfo, Word& w )
   else
   {
     R_Tree<dim, TupleId>* rtree = (R_Tree<dim, TupleId>*)w.addr;
-    rtree->DeleteFile();
-    delete rtree;
+/*    rtree->DeleteFile();
+    delete rtree;*/
+    if(rtree->GetShare() == 1){
+      rtree->CloseFile();
+    }else{
+      rtree->DeleteFile();
+      delete rtree;
+     }
   }
 }
